Fix the UI of AddPostActivity and Refactor package
- add addPost() on Book model 
- add textView showing current date
- refactor utils to util package
